stdlib: mark Float.value as private
Swift SVN r20590

+++ test/Interpreter/builtin.swift
@@ -1,8 +1,10 @@
-// RUN: %target-run-stdlib-swift | FileCheck %s
+// RUN: mkdir -p %t
+// RUN: %target-build-swift %s -parse-stdlib -Xfrontend -disable-access-control -o %t/a.out
+// RUN: %target-run %t/a.out | FileCheck %s
 
 import Swift
 
-println(String(UInt64(Int32(Builtin.bitcast_FPIEEE32_Int32(Float32(1).value))), radix: 16)) // CHECK: {{^}}3f800000{{$}}
+println(String(Int32(Builtin.bitcast_FPIEEE32_Int32(Float32(1).value)), radix: 16)) // CHECK: {{^}}3f800000{{$}}
 println(String(UInt64(Builtin.bitcast_FPIEEE64_Int64(Float64(1).value)), radix: 16)) // CHECK: {{^}}3ff0000000000000{{$}}
 
 // Check that floating point literals work
